fix: move types folder
