+++ src/app/patient-summary/components/resources/Organization/Organization.tsx
@@ -2,7 +2,7 @@ import { StyledSectionTypography } from '@/app/components/typography/StyledTypog
 import InfoRow from '../../InfoRow';
 import ConnectionDetails from '../Patient/ConnectionDetails';
 import React from 'react';
-import { IResourceType } from '@/app/patient-summary/types/fhir.types';
+import { IResourceType } from '@/types/fhir.types';
 
 export default function Organization({
   data,
