+++ src/app/utils/helpers.tsx
@@ -1,8 +1,4 @@
-import {
-  TType,
-  TBundle,
-  IDynamicProps,
-} from '../patient-summary/types/fhir.types';
+import { TType, TBundle, IDynamicProps } from '../../types/fhir.types';
 
 export const extractResourceInfo = <
   TResource extends IDynamicProps['resource'],
